Project/python/unet-lightning-master/requirements.txt
absl-py==0.10.0
alembic==1.4.3
appdirs==1.4.4
astunparse==1.6.3
cachetools==4.1.1
camera==1.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataset==1.3.2
decorator==4.4.2
distlib==0.3.1
docutils==0.16
filelock==3.0.12
fsspec==0.8.3
future==0.18.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
joblib==1.0.0
Keras-Preprocessing==1.1.2
lxml==4.6.2
Mako==1.1.3
Markdown==3.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
nltk==3.5
nn==0.1.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.4
Pillow==7.2.0
plyer==2.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.3
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-lightning==0.8.5
pytorch-model-summary==0.1.2
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
PyYAML==5.3.1
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.24.0
scipy==1.5.4
self==2020.12.3
six==1.15.0
SQLAlchemy==1.3.19
summary==0.2.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.11.18
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
urllib3==1.25.10
virtualenv==20.4.3
Werkzeug==1.0.1
wrapt==1.12.1
